Separated mockDebounce to new spec-helper-functions.js and added more comments
